# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaptureAnalysisInputSource = typeof CaptureAnalysisInputSource[keyof typeof CaptureAnalysisInputSource];


export const CaptureAnalysisInputSource = {
  note: 'note',
  photo: 'photo',
  voice: 'voice',
} as const;

export interface CaptureAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
  source: CaptureAnalysisInputSource;
}

export type CaptureAnalysisCategory = typeof CaptureAnalysisCategory[keyof typeof CaptureAnalysisCategory];


export const CaptureAnalysisCategory = {
  personal: 'personal',
  practical: 'practical',
  people: 'people',
  travel: 'travel',
  work: 'work',
  object: 'object',
  other: 'other',
} as const;

export interface CaptureAnalysis {
  signal: string;
  likelyOmission: string;
  explanation: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  preventiveAction: string;
  category: CaptureAnalysisCategory;
}