# lib/api-zod/src/generated/types/captureAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaptureAnalysisCategory } from './captureAnalysisCategory';

export interface CaptureAnalysis {
  signal: string;
  likelyOmission: string;
  explanation: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  preventiveAction: string;
  category: CaptureAnalysisCategory;
}