# lib/api-zod/src/generated/types/captureAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaptureAnalysisInputSource } from './captureAnalysisInputSource';

export interface CaptureAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
  source: CaptureAnalysisInputSource;
}